solutions week2 e04
