+++ week-2-wide-perspective/w02e04-dungeons-and-the-princess/package.json
@@ -1,5 +1,15 @@
 {
   "name": "w02e04-dungeons-and-the-princess",
   "version": "1.0.0",
-  "dependencies": {}
+  "scripts": {
+    "dev": "nodemon src/main.ts"
+  },
+  "dependencies": {
+    "nodemon": "2.0.20",
+    "ts-node": "10.9.1"
+  },
+  "devDependencies": {
+    "@types/node": "*",
+    "typescript": "4.8.4"
+  }
 }
