+++ week-2-wide-perspective/w02e04-dungeons-and-the-princess/src/Dungeon.ts
@@ -1,0 +1,14 @@
+import { Underground } from './Underground';
+import { Person } from './Person';
+
+export class Dungeon {
+  constructor(private underground: Underground) {}
+
+  enter(p: Person) {
+    if (p.title === "queen") {
+      throw new Error("Queen cannot enter the dungeon!")
+    } else {
+      this.underground.enter(p)
+    }
+  }
+}
