+++ week-2-wide-perspective/w02e04-dungeons-and-the-princess/src/Person.ts
@@ -1,0 +1,13 @@
+import { Title } from './Title';
+
+export class Person {
+  constructor(private name: string, public readonly title: Title) {}
+
+  shoutTheName() {
+    console.log(`I'am ${this.name}! The ${this.title}.`)
+  }
+
+  sayThanks() {
+    console.log(`${this.name}: Thank you...`)
+  }
+}
