+++ week-2-wide-perspective/w02e04-dungeons-and-the-princess/src/Title.ts
@@ -1,0 +1,1 @@
+export type Title = 'princess' | 'king' | 'queen' | 'knight' | 'peasant'
